Avoid a circular import
tasks/__init__ -> views/site_evaluation -> views/__init__ ->
views/vector_tile -> views/model_run -> tasks/__init__
